Added an option for a progress callback function
Provides an integer percent of the upload based on the content-length
and upload totalbytes. It is not exact, but it's good enough simple
indication. Is called once on end with value of 100. (note using the
server.upload() or Update.status global functions in the callback allows
more depth of what's going on, didn't seem neccesary to pass them.)